Updates from Day 1
I was able to successfully create and render a Portfolio page and link
to my existing Github repos.  As a next step with that, I'd want to
embed some screenshots and images as a reference point to said projects.
Also, I started the process of setting up some bones for a Contact form
and got it to render but not successfully connect.  That's tomorrow's
adventure.