Initial Styling
For todays successes, I was able to successfully embed some
screenshots of the loading page for each of my GA apps.  Also, I began
the process of basic styling.

In terms of adjustments, I added my contact info to the footer so it's
viewable on each page and removed the Contact page.  I also added a
Home component with my About info.  For a next step, I need to work on
adding my resume and I'd be interested in seeing if I can link or
embed my LinkedIn as perhaps a widget.